[qira] move everything to qira.c from tci.c
